Сделана основа блока уведомлений, views разнесено по файлам
